+++ catch_client/src/systems/draw_item_system.rs
@@ -1,4 +1,4 @@
-use std::f64;
+use std::f32;
 
 use ecs::{Aspect, System, DataHelper, Process};
 
@@ -25,12 +25,12 @@ impl DrawItemSystem {
         }
     }
 
-    pub fn draw(&mut self, data: &mut DataHelper<Components, Services>, time_s: f64,
+    pub fn draw(&mut self, data: &mut DataHelper<Components, Services>, time_s: f32,
                 particles: &mut Particles, c: graphics::Context, gl: &mut GlGraphics) {
         for entity in self.aspect.iter() {
             let p = data.position[entity].p;
             let size = match data.shape[entity] {
-                Shape::Square { size } => size,
+                Shape::Square { size } => size as f64,
                 _ => panic!("item should be square"),
             };
 
@@ -41,16 +41,16 @@ impl DrawItemSystem {
                                      color, color,
                                      1.5, // size
                                      p, // position
-                                     data.orientation[entity].angle - f64::consts::PI,
-                                     f64::consts::PI * 2.0,
-                                     20.0 + rand::random::<f64>() * 10.0, // speed
-                                     f64::consts::PI,
+                                     data.orientation[entity].angle - f32::consts::PI,
+                                     f32::consts::PI * 2.0,
+                                     20.0 + rand::random::<f32>() * 10.0, // speed
+                                     f32::consts::PI,
                                      0.0 // friction
                                      );
             }
 
-            let transform = c.trans(p[0], p[1])
-                             .rot_rad(data.orientation[entity].angle)
+            let transform = c.trans(p[0] as f64, p[1] as f64)
+                             .rot_rad(data.orientation[entity].angle as f64)
                              .transform;
 
             graphics::rectangle([0.0, 1.0, 0.0, 1.0],
