+++ catch_client/src/systems/draw_projectile_system.rs
@@ -22,18 +22,19 @@ impl DrawProjectileSystem {
         }
     }
 
-    pub fn draw(&mut self, data: &mut DataHelper<Components, Services>, time_s: f64,
+    pub fn draw(&mut self, data: &mut DataHelper<Components, Services>, time_s: f32,
                 particles: &mut Particles, c: graphics::Context, gl: &mut GlGraphics) {
         for entity in self.aspect.iter() {
             let p = data.position[entity].p;
             let angle = data.orientation[entity].angle;
 
             let (w, h) = match data.shape[entity] {
-                Shape::Rect { width, height } => (width, height),
+                Shape::Rect { width, height } => (width as f64, height as f64),
                 _ => panic!("projectile should be rect"),
             };
 
-            let transform = c.trans(p[0], p[1]).rot_rad(angle).transform;
+            let transform = c.trans(p[0] as f64, p[1] as f64)
+                             .rot_rad(angle as f64).transform;
 
             graphics::rectangle([0.2, 0.2, 0.2, 1.0],
                                 [-w/2.0, -h/2.0, w, h],
